+++ packages/veui/demo/cases/NumberInput.vue
@@ -21,11 +21,14 @@
         <veui-field label="DecimalPlace = -1：" tip="不处理精度问题">
           <veui-number-input v-model="number5" :decimal-place="-1"/>
         </veui-field>
+        <veui-field label="Max &amp;&amp; Min：" :rules="[{name: 'min', value: -1}]" tip="最大值不大于 10，最小值不小于 -1">
+          <veui-number-input v-model="number6" :max="10"/>
+        </veui-field>
         <veui-field label="Readonly：">
-          <veui-number-input readonly v-model="number6"/>
+          <veui-number-input readonly v-model="number7"/>
         </veui-field>
         <veui-field label="Disabled：" disabled>
-          <veui-number-input v-model="number6"/>
+          <veui-number-input v-model="number7"/>
         </veui-field>
       </veui-form>
     </section>
@@ -50,7 +53,8 @@ export default {
       number3: null,
       number4: null,
       number5: null,
-      number6: 1024
+      number6: null,
+      number7: 1024
     }
   }
 }